Update pinned dep versions, README (#110)
* Update pinned image dependencies
* Update README to reflect latest release
* Add additional note to README about working around broken builds (b/c of pinned versions)

The security update involves libtiff. The exploit requires using a crafted TIFF image. This shouldn't be a problem for us because we control the TIFF images served through Cantaloupe, so we're not creating a new Docker release at this time. We do want to update the version, though, so that the build runs without errors.