Update patch
When the HTTP response code for a resource will be 401, the patched
method should return true if the resource is an info.json, so that
Cantaloupe will return JSON-LD, rather than an HTML error page with a
stack trace. HTTP 401 responses for image requests, however, should
still render an error HTML page (and thus the method in question should
still throw an exception in that case).